requirements.txt
allure-pytest
pytest
fastapi
jinja2
python-multipart
httpx
